split, add currentConfig to each function on the meta
